Add support for the new voice record/playback feature
The upload to the LMS is done using a smb share.
Two new entries are added to the config.json to configure this.
The uploadPath path is the location where the saved
file should be stored and can be a smb mount to the LMS server.
The uploadPlaybackPath is the location on the smb and is used
to start the playback.