Call scripts for all actions which can't be done in the MSG
E.g. for changing the equalizer, or playing the uploaded voice
recording.

The scripts are configured using the "scripts" section in the
config.json. Every command has it's own entry in the config.json,
but this doesn't mean it also needs to be separate script which
is called.
Every script can use the provided {{arg}} placeholders in their
commands. These entries will be replaced by the MSG before
executing the command.