lms: Fix saving the correct URL in (zone)favorites
For most operations the id of the items is enough, but for storing
them in the LMS favorite list or in the zone favorites it isn't
as the LMS doesn't support our id schema and in zone favorites
it's also not a good idea as those ids change when the DB is
reindexed.

The change introduces function which resolves the URL for
the passed items. For some this is easy, but for artists/albums...
we are using a LMS internal schema which is also understood by
the LMS favorites.

Retrieving the URLs from the favorite list is also crucial and
until https://github.com/Logitech/slimserver/issues/459 is fixed
Storing Albums/Artist/... works fine, but restoring them from the
zone favorites might not play the correct item.