Disable the list cache
The cache is only useful for really long lists e.g. the library.
For some other lists e.g. radio and spotify it is not very convenient
as it needs to be invalidated to show uptodate data.

For now we disable the cache until we know it is really needed.