Add support for zone syncing
This includes create groups, adding/removing zones, deleting the
group completely.

The current track/playlist is kept in sync between all those zones
in a group. Once a zone is powered off it is still kept in sync,
but it will show it as not playing.
Once the zone is powered on again, it continues playing in sync
with the other group members and picks up the playback.

When powering off the master zone of a group, the master zone
changes to be the next group member.