Add a workaround for the slow image handling of the app
This workaround provides a image only for the current Item in
the queue. On every update of the current track a queue_event is
send to let the app fetch the new queue and get the new image.

This is needed as otherwise the app needs very long to delete
all old images and fetch the new images and is unresponsive while
doing this.