Add code to get informed about Spotify account changes in the app
The app doesn't notify the MSG about a change in the usual way and
we also don't have a way to know which zone currently talks to us.

This change introduces a dependency to LxCommunicator, which is
needed for the communication to the Miniserver.

We get notified once the stored settings on the MS changes. This
triggers a change to the same user account on this zone, using
the matching spotty commands.

The current spotify account name is also saved in each zone, as
this is needed to show the correct data in the servicefolder command.

As we always want to show the content for the current User, in current
zone. We need to use the correct zones mac for the spotty commands.
As we also don't get the zone information in the servicefolder command
we search all zones for one with the requested user and will use this
mac for showing the content.

TODO:
  * Add a better password storage, currently it needs to be stored in
    plain text.
  * Add logging categories for the ms client
  * Add reconnection handling for the ms client