Add syncing of usersettings over multiple users
The usersettings change notification also contains the information
for which user the settings have changed. Retrieving the user setting
for another user is not possible, so we have to maintain a list
of users we want to sync. The config.json has been adapted accordingly.

Whenever one of the settings of our user changes, the current active
accounts are also written to the settings of all other users to
keep them in sync and always show the correct active user of the zone.