Add Audioserver Zone Mapping
The playerids for all configured Zones are not static like they
were on the Musicserver. Because of that the zones are now read
matched by name. To support this, the internal zone list needed
to be changed to a Map and only the needed zones return valid
replies, while all other reply some dummy values.