Add support for the diagnosis command
Similar to other commands this command will call a bash script
configured in the config.json.

The executed script can return a string which is interpreted as
JSON and shown in the app.

Also improve the error handling when not all scripts are configured
and a empty string is returned.